requirements.txt
fastapi
uvicorn
aiofile
pyjwt
